+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 
 ### Added
 
-- Add `hbase.rest.hostname`, `hbase.rest.port`, and `hbase.rest.info.port` properties to the restserver `hbase-site.xml` ([#708]).
+- Add custom `hbase.rest.endpoint` property to the restserver `hbase-site.xml` ([#708], [#716]).
+  - The custom `hbase.rest.hostname` and native `hbase.rest.port` properties cannot be used for
+    discovery advertisement, as these should remain stable regardless of listener class used to expose the REST service.
 
 [#708]: https://github.com/stackabletech/hbase-operator/pull/708
+[#716]: https://github.com/stackabletech/hbase-operator/pull/716
 
 ## [25.11.0] - 2025-11-07